Intern glyph names
This began with me just wanting to change Component to use
GlyphName instead of String, but it felt bad that we were going to
be storing multiple copies of the names, so I also implemented
a basic interning scheme; while parsing we pass along a HashSet of
names; before using a name, we check that it doesn't already exist in
the map.

When we implement parallel loading we will need to put this HashSet
behind a RwLock or similar, but that shouldn't be too difficult.