Add postscript glyph name lookup
This uses a buildscript to generate a perfect hashmap for the glyph
names. This is maybe a bit over the top; we could have just put them
in order and done a binary search?

But we can always change this down the road if it feels like we're
definitely being too fancy?