Add LayerSet, other related API changes
This adds a LayerSet type that enforces various constraints on
layers.

Among other things, this means that:

- Layer names are now also stored as Arc<str>
- There is no more LayerInfo type
- there is now always a default layer
- the default layer cannot be removed
- new layers can be added

This also introduces a few other breaking api changes; for instance
I'm removing `Font::get_default_layer(&self) -> Option<&Layer>`
in favour of `Font::default_layer(&self) -> &Layer`.

This was motivated by my experiments with python bindings,
but I think this is generally useful independent of that work
and so I'm PRing it separately.