Rework loading methods
This is mainly intended to clear up confusion around how to use
load_ufo; the answer is "don't".

I believe that the previous API of splitting up the creation of the font
from the loading was unnecessary. I also suspect that having the Font
object hold on to the DataRequest is unnecessary, although there are
possible arguments in favor of this approach.

In any case, the main confusion with the previous API is that loading
with a data request took `&self` by reference but returned a new font
object, which is confusing. I would expect it either to take a
DataRequest as an argument to an associated function, or I would expect
it to take `&mut self` and load the contents of the file into the
current font?