Add Name type
Glyph and layer names cannot be arbitary strings. This adds a name
module and a Name type that handles validation.

This validation is always performed when loading a font. At runtime, we
expect the user to ensure that they are using a valid name, since
invalid names are generally hard to construct.