Totally rework .glif version handling
This is a simplification that better reflects the crate's actual
behaviour.

- we now parse versionMinor field, if present
- glif version is now an internal detail. When we read a glif, we verify
that the version is one that we support, and we keep track of the
version during parsing, as it impacts parsing behaviour; but once
parsing has succeeded we discard this information. When we serialize, we
always serialize to version (2, 0).
- the user can no longer provide a .glif version when creating a Glyph.
- we no longer need to check the version when serializing, since only
  version 2 compatible glyphs are constructable.