Revert API change in 53c603c
In the previous version of quick-xml there was an issue that prevented
deserialization of Option<Vec<_>>. We worked around this by changing our
API, but that change was not an improvement, and the issue is fixed in
quick-xml 0.28. Since I intend to yank norad 0.9, we can revert that
change, providing a better upgrade path for our users.