Make uuid an optional dependency
We use the uuid crate to generate object identifiers, and we use object
identifiers anytime we set a lib object on a type where the lib is
stored in the parent (such as contour or guideline).

This adds a new feature, `object-libs`, that brings in `uuid` and gates
all the API that is involved in setting the libs on these objects.

This also changes various constructors to not take a `lib` directly, and
instead the lib will have to be manually added by the user.